Retrieve order and its items from the database
In order to retrieve the order and its items
from the database I should have used the `as`("items") and then convertFrom.
Refactor order repository to use long instead of
int.

+++ src/main/kotlin/com/example/dao/Orders.kt
@@ -7,8 +7,8 @@ import com.example.models.OrderItemCreation
 
 interface Orders {
     suspend fun getAll() : List<OrderCreated>
-    suspend fun get(id: Int?): OrderCreated?
-    suspend fun add(order: OrderCreation): Int?
-    suspend fun addOrderItem(orderId: Int, item: OrderItemCreation): Int
-    suspend fun delete(id: Int?): Boolean
+    suspend fun get(id: Long?): OrderCreated?
+    suspend fun add(order: OrderCreation): Long?
+    suspend fun addOrderItem(orderId: Long, item: OrderItemCreation): Long
+    suspend fun delete(id: Long?): Boolean
 }
